+++ index.js
@@ -7,95 +7,141 @@
  */
 
 var http = require('http');
-var XHR = require('xmlhttprequest-socket').XMLHttpRequest;
 
-function eventHandler(sockPath, server) {
-  this.sockPath = sockPath;
-  this.server = server;
+function eventHandler(application, listeningCallback) {
+  this.application = application;
   this.serverReady = false;
+  this.socketPathSuffix = 0;
+  this.server = this.createServer(application, listeningCallback);
 };
 
-eventHandler.prototype.handle = function(event, context, callback) {
+eventHandler.prototype.createServer = function(application, listeningCallback) {
   var self = this;
-  this.event = event;
-  this.context = context;
-  if (this.serverReady) {
-    this.sendRequest(callback);
-  }
-  else {
-    this.server.on('listening', function() {
-      self.sendRequest(callback);
-    });
-  }
+  var server = http.createServer(application);
+  server.on('listening', function() {
+    self.serverReady = true;
+    if (listeningCallback) {
+      listeningCallback(self.getSocketPath());
+    }
+  }).on('close', function() {
+    self.serverReady = false;
+  }).on('error', function(error) {
+    if (error.code === 'EADDRINUSE') {
+      console.warn(`EADDRINUSE ${self.getSocketPath()} incrementing socketPathSuffix.`);
+      ++self.socketPathSuffix;
+      self.server.close(function() {
+        self.startServer();
+      });
+    }
+  });
+  return server;
 };
 
-eventHandler.prototype.sendRequest = function(callback) {
-  var handled = false;
-  var request = new XHR();
-  request.onreadystatechange = function() {
-    var isJson;
-    if (request.readyState === request.HEADERS_RECEIVED && !handled) {
-      var contentType = request.getResponseHeader('Content-Type');
-      isJson = contentType && contentType.indexOf('application/json') === 0;
-    }
-    if (request.readyState !== request.DONE || handled) {
-      return;
-    }
-    handled = true;
-    if (request.status === 0) {
-        callback('Unable to connect to the specified socket');
-    }
-    else {
-      var response;
+eventHandler.prototype.returnResponseFromLambda = function(response, context, callback) {
+  var chunks = '';
+  response.setEncoding('utf8');
+  response.on('data', function(chunk) {
+    chunks+= chunk.toString('utf8');
+  }).on('end', function() {
+    var statusCode = response.statusCode;
+    var headers = response.headers;
+    Object.keys(headers).forEach(function(h) {
+      if (Array.isArray(headers[h])) {
+        headers[h] = headers[h].join(',');
+      }
+    });
+    var body = '';
       try {
-        response = JSON.parse(request.responseText);
+        body = JSON.parse(chunks);
       }
       catch (error) {
-        response = request.responseText;
+        body = chunks;
       }
-      if (request.status >= 200 && request.status < 300) {
-        callback(null, response);
-      }
-      else {
-        if (typeof response === 'object') {
-          response = JSON.stringify(response)
-        }
-        callback(response);
-      }
-    }
-    }
-  request.openOnSocket(this.event.method, this.event.path, this.sockPath, true);
-  for (var h in this.event.headers) {
-    request.setRequestHeader(h, this.event.headers[h]);
+    var successResponse = {statusCode, headers, body};
+
+    callback(null, successResponse);
+  })
+}
+
+eventHandler.prototype.connectionError = function(error, context, callback) {
+  var errorResponse = {
+    statusCode: 502,
+    body: '',
+    headers: {}
   }
-  var data;
-  if (this.event.body) {
-    if (typeof this.event.body ==='string') {
-      data = this.event.body;
+
+  callback(errorResponse);
+}
+
+eventHandler.prototype.forwardRequestToLambda = function(event, context, callback) {
+  var headers = event.headers || {};
+  var requestOptions = {
+    method: event.httpMethod || event.method || 'GET',
+    path: event.path,
+    headers: headers,
+    socketPath: this.getSocketPath()
+  };
+  var self = this;
+  var request = http.request(requestOptions, function(response) {
+    self.returnResponseFromLambda(response, context, callback);
+  });
+
+  if (event.body) {
+    request.write(JSON.stringify(event.body));
+  }
+
+  request.on('error', function(error) {
+    self.connectionError(error, context, callback);
+  }).end();
+}
+
+eventHandler.prototype.internalServerError = function(error, context, callback) {
+    var errorResponse = {
+        statusCode: 500,
+        body: '',
+        headers: {}
+    }
+
+    callback(errorResponse);
+}
+
+eventHandler.prototype.getSocketPath = function() {
+    return `/tmp/rooftop${this.socketPathSuffix}.sock`
+}
+
+eventHandler.prototype.startServer = function() {
+    return this.server.listen(this.getSocketPath());
+}
+
+eventHandler.prototype.handleEvent = function(event, context, callback) {
+  try {
+    if (this.serverReady) {
+      this.forwardRequestToLambda(event, context, callback);
     }
     else {
-      data = JSON.stringify(this.event.body);
+      var self = this;
+      this.startServer();
+      this.server.on('listening', function() {
+        try {
+          self.forwardRequestToLambda(event, context, callback);
+        }
+        catch (error) {
+          self.internalServerError(error, context, callback);
+        }
+      });
     }
   }
-  request.send(data);
+  catch (error) {
+    this.internalServerError(error, context, callback);
+  }
 };
 
-var lampress = function(sockPath, getServer) {
-  var server;
-  var handler;
+var lampress = function(application, listeningCallback) {
+  var handler = new eventHandler(application, listeningCallback);
   return function(event, context, callback) {
     context.callbackWaitsForEmptyEventLoop = false;
     global.Context = context;
-    if (!server) {
-      server = getServer();
-    }
-    if (!handler) {
-      handler = new eventHandler(sockPath, server);
-      server.on('listening', function() {
-        handler.serverReady = true;
-      });
-    }
-    handler.handle(event, context, callback);
+    handler.handleEvent(event, context, callback);
   };
 };