+++ HISTORY.md
@@ -1,3 +1,13 @@
+2.0.0 / 2016-10-24
+===================
+
+  * Full rewrite to use Node's own request
+
+1.1.1 / 2016-07.25
+===================
+
+  * Make callback a parameter of sendRequest
+
 1.1.0 / 2016-07.23
 ===================